+++ packages/redux/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Next
 
+* Switch to `router` from `config`.
 * Removed source maps from `dist`
 
 ## 1.0.0-beta.1